+++ src/internal/nft-api.ts
@@ -5,7 +5,10 @@ import {
   GetFloorPriceResponse,
   GetNftsForContractOptions,
   GetNftsForOwnerOptions,
+  GetOwnersForContractOptions,
   GetOwnersForContractResponse,
+  GetOwnersForContractWithTokenBalancesOptions,
+  GetOwnersForContractWithTokenBalancesResponse,
   GetOwnersForNftResponse,
   NftContractBaseNftsResponse,
   NftContractNftsResponse,
@@ -199,17 +202,27 @@ export async function* getNftsForContractIterator(
 export async function getOwnersForContract(
   config: AlchemyConfig,
   contractAddress: string,
+  options?:
+    | GetOwnersForContractWithTokenBalancesOptions
+    | GetOwnersForContractOptions,
   srcMethod = 'getOwnersForContract'
-): Promise<GetOwnersForContractResponse> {
-  const response = await requestHttpWithBackoff<
+): Promise<
+  GetOwnersForContractResponse | GetOwnersForContractWithTokenBalancesResponse
+> {
+  // Cast to `any` to avoid more type wrangling.
+  const response: any = await requestHttpWithBackoff<
     GetOwnersForNftContractAlchemyParams,
     RawGetOwnersForContractResponse
   >(config, AlchemyApiType.NFT, 'getOwnersForCollection', srcMethod, {
+    ...options,
     contractAddress
   });
 
   return {
-    owners: response.ownerAddresses
+    owners: response.ownerAddresses,
+
+    // Only include the pageKey in the final response if it's defined
+    ...(response.pageKey !== undefined && { pageKey: response.pageKey })
   };
 }
 
@@ -250,6 +263,53 @@ export async function checkNftOwnership(
     srcMethod
   );
   return response.ownedNfts.length > 0;
+}
+
+export async function verifyNftOwnership(
+  config: AlchemyConfig,
+  owner: string,
+  contractAddresses: string | string[],
+  srcMethod = 'verifyNftOwnership'
+): Promise<boolean | { [contractAddress: string]: boolean }> {
+  if (typeof contractAddresses === 'string') {
+    const response = await getNftsForOwner(
+      config,
+      owner,
+      {
+        contractAddresses: [contractAddresses],
+        omitMetadata: true
+      },
+      srcMethod
+    );
+    return response.ownedNfts.length > 0;
+  } else {
+    if (contractAddresses.length === 0) {
+      throw new Error('Must provide at least one contract address');
+    }
+    const response = await getNftsForOwner(
+      config,
+      owner,
+      {
+        contractAddresses,
+        omitMetadata: true
+      },
+      srcMethod
+    );
+
+    // Create map where all input contract addresses are set to false, then flip
+    // owned nfts to true.
+    const result = contractAddresses.reduce(
+      (acc: { [contractAddress: string]: boolean }, curr) => {
+        acc[curr] = false;
+        return acc;
+      },
+      {}
+    );
+    for (const nft of response.ownedNfts) {
+      result[nft.contract.address] = true;
+    }
+    return result;
+  }
 }
 
 export async function isSpamContract(